refactor(shared): rename protocol.js to MessageProtocol.js to match class name
