+++ docs/hetic/stack.md
@@ -14,4 +14,4 @@
 
 - **Cannon-es** (physique, backend initial) : utilisé jusqu'à la migration Rapier. Retiré une fois Rapier validé pour profiter d'un solver plus moderne, des `EventQueue` natifs et du support CCD potentiel. Le port `adapters/physics/ports/PhysicsPort.js` permet de revenir à Cannon-es si besoin.
 - **Ammo.js** (physique) : plus lourd (~1,5 MB) ; Rapier retenu pour son API moderne et son ecosystème Rust/WASM.
-- **Socket.IO** (communication, transport initial) : utilisé jusqu'à la migration vers **WebSocket natif (`ws`)**. Retiré pour supprimer une dépendance lourde ; la reconnexion automatique et le multiplexage d'événements nommés sont désormais assurés par `shared/src/realtimeClient.js` (façade `on`/`off`/`emit`) et `shared/src/protocol.js` (codec d'enveloppe `{ event, data }`), le broadcast par `server/src/adapters/transport/GameBroadcaster.js`.
+- **Socket.IO** (communication, transport initial) : utilisé jusqu'à la migration vers **WebSocket natif (`ws`)**. Retiré pour supprimer une dépendance lourde ; la reconnexion automatique et le multiplexage d'événements nommés sont désormais assurés par `shared/src/realtimeClient.js` (façade `on`/`off`/`emit`) et `shared/src/MessageProtocol.js` (codec d'enveloppe `{ event, data }`), le broadcast par `server/src/adapters/transport/GameBroadcaster.js`.
