+++ shared/src/MessageProtocol.js
