+++ shared/src/index.js
@@ -1,3 +1,3 @@
 export * from "./eventNames.js";
-export * from "./protocol.js";
+export * from "./MessageProtocol.js";
 export * from "./realtimeClient.js";
